have ladder climbing working, movement script is good, just need to have the player go through the platform so that they can actually get up to the next one.
