+++ ProjectSettings/TagManager.asset
@@ -16,9 +16,9 @@ TagManager:
   - 
   - Water
   - UI
-  - 
-  - 
-  - 
+  - Climbing
+  - Player
+  - Platform
   - 
   - 
   -
